第一版完成

+++ src/main/java/com/by/blcu/mall/controller/CourseCategoryInfoController.java
@@ -1,8 +1,11 @@
 package com.by.blcu.mall.controller;
 
+import com.alibaba.fastjson.JSONObject;
+import com.by.blcu.core.aop.CheckToken;
 import com.by.blcu.core.ret.RetResult;
 import com.by.blcu.core.ret.RetResponse;
 import com.by.blcu.core.utils.ApplicationUtils;
+import com.by.blcu.core.utils.StringUtils;
 import com.by.blcu.mall.model.CourseCategoryInfo;
 import com.by.blcu.mall.service.CourseCategoryInfoService;
 import com.by.blcu.mall.vo.CourseCategoryInfoVo;
@@ -14,8 +17,10 @@ import io.swagger.annotations.ApiParam;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 /**
 * @Description: CourseCategoryInfoController类
@@ -32,8 +37,8 @@ import java.util.List;
         "5、课程分类下移接口\n" +
         "6、根据父ID查询接口\n" +
         "7、递归查询所有接口")
+@CheckToken
 public class CourseCategoryInfoController {
-
     @Resource
     private CourseCategoryInfoService courseCategoryInfoService;
 
@@ -46,22 +51,28 @@ public class CourseCategoryInfoController {
 
     @PostMapping("/insertByParentId")
     @ApiOperation(value = "添加课程分类接口")
-    public RetResult<Integer> insertByParentId(CourseCategoryInfo courseCategoryInfo) throws Exception{
+    public RetResult<Integer> insertByParentId(@RequestBody CourseCategoryInfo courseCategoryInfo,HttpServletRequest httpServletRequest) throws Exception{
         courseCategoryInfo.setCcId(ApplicationUtils.getUUID());
         String parentId = courseCategoryInfo.getParentId();
-        if(parentId != null && !"".equals(parentId)){
+        String ccName = courseCategoryInfo.getCcName();
+        List<CourseCategoryInfo> courseCategoryInfos = courseCategoryInfoService.selectByParentIdAndName(parentId,ccName.trim());
+        if(null != courseCategoryInfos && courseCategoryInfos.size()>0){
+            return RetResponse.makeErrRsp("该分类名已存在！");
+        }
+        if(parentId != null && !"".equals(parentId) && ccName != null && !"".equals(ccName)){
             Integer sort = courseCategoryInfoService.selectMaxSortByParentId(parentId);
             if(sort == null){
                 courseCategoryInfo.setCcSort(0);
             }else{
                 courseCategoryInfo.setCcSort(sort+1);
             }
+            courseCategoryInfo.setCcCreator((String) httpServletRequest.getAttribute("username"));
             courseCategoryInfo.setCcStatus(1);
             courseCategoryInfo.setCcCreateTime(new Date());
             Integer state = courseCategoryInfoService.insert(courseCategoryInfo);
             return RetResponse.makeOKRsp(state);
         } else {
-            return RetResponse.makeErrRsp("parentId不能为空！");
+            return RetResponse.makeErrRsp("parentId和ccName不能为空！");
         }
     }
 
@@ -73,9 +84,14 @@ public class CourseCategoryInfoController {
 
     @PostMapping("/deleteByCcId")
     @ApiOperation(value = "递归删除接口")
-    public RetResult<Integer> deleteByCcId(@RequestParam String ccId) throws Exception {
-        Integer state = courseCategoryInfoService.deleteByCcId(ccId);
-        return RetResponse.makeOKRsp(state);
+    public RetResult<Integer> deleteByCcId(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("ccId") && !StringUtils.isBlank(obj.getString("ccId"))){
+            String ccId = obj.getString("ccId");
+            Integer state = courseCategoryInfoService.deleteByCcId(ccId);
+            return RetResponse.makeOKRsp(state);
+        }else{
+            return RetResponse.makeErrRsp("需要选择要删除的分类！");
+        }
     }
 
     @PostMapping("/update")
@@ -86,30 +102,54 @@ public class CourseCategoryInfoController {
 
     @PostMapping("/updateNameByCcId")
     @ApiOperation(value = "编辑名字接口")
-    public RetResult<Integer> updateNameByCcId(@RequestParam String ccId,@RequestParam String ccName) throws Exception {
-        Integer state = courseCategoryInfoService.updateNameByCcId(ccId,ccName);
-        return RetResponse.makeOKRsp(state);
+    public RetResult<Integer> updateNameByCcId(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("ccId") && obj.containsKey("ccName") && !StringUtils.isBlank(obj.getString("ccId")) && !StringUtils.isBlank(obj.getString("ccName"))){
+            String ccId = obj.getString("ccId");
+            String ccName = obj.getString("ccName");
+            CourseCategoryInfo courseCategoryInfo = courseCategoryInfoService.selectByCcId(ccId);
+            if(null == courseCategoryInfo){
+                return RetResponse.makeErrRsp("不存在该分类！");
+            }
+            List<CourseCategoryInfo> courseCategoryInfos = courseCategoryInfoService.selectByParentIdAndName(courseCategoryInfo.getParentId(),ccName);
+            if(null != courseCategoryInfos && courseCategoryInfos.size()>0){
+                return RetResponse.makeErrRsp("该分类名已存在！");
+            }
+            Integer state = courseCategoryInfoService.updateNameByCcId(ccId,ccName);
+            return RetResponse.makeOKRsp(state);
+        }else{
+            return RetResponse.makeErrRsp("需要传入参数ccId和ccName！");
+        }
     }
 
     @PostMapping("/updateUpSort")
     @ApiOperation(value = "课程分类上移接口")
-    public RetResult<Integer> updateSort(String ccId) throws Exception {
-        Integer state = courseCategoryInfoService.updateUpSort(ccId);
-        if (state == 1){
-            return RetResponse.makeOKRsp(state);
-        } else{
-            return RetResponse.makeErrRsp("已经置顶！");
+    public RetResult<Integer> updateSort(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("ccId")){
+            String ccId = obj.getString("ccId");
+            Integer state = courseCategoryInfoService.updateUpSort(ccId);
+            if (state == 1){
+                return RetResponse.makeOKRsp(state);
+            } else{
+                return RetResponse.makeErrRsp("已经置顶！");
+            }
+        }else{
+            return RetResponse.makeErrRsp("需要传入参数ccId！");
         }
     }
 
     @PostMapping("/updateDownSort")
     @ApiOperation(value = "课程分类下移接口")
-    public RetResult<Integer> updateDownSort(String ccId) throws Exception {
-        Integer state = courseCategoryInfoService.updateDownSort(ccId);
-        if (state == 1){
-            return RetResponse.makeOKRsp(state);
-        } else{
-            return RetResponse.makeErrRsp("已经到底！");
+    public RetResult<Integer> updateDownSort(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("ccId")){
+            String ccId = obj.getString("ccId");
+            Integer state = courseCategoryInfoService.updateDownSort(ccId);
+            if (state == 1){
+                return RetResponse.makeOKRsp(state);
+            } else{
+                return RetResponse.makeErrRsp("已经到底！");
+            }
+        }else{
+            return RetResponse.makeErrRsp("需要传入参数ccId！");
         }
     }
 
@@ -119,13 +159,39 @@ public class CourseCategoryInfoController {
         return RetResponse.makeOKRsp(courseCategoryInfo);
     }
 
+    @PostMapping("/selectCcNameByCcId")
+    public RetResult<String> selectCcNameByCcId(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("ccId")){
+            String ccId = obj.getString("ccId");
+            String ccName = courseCategoryInfoService.selectCcNameByCcId(ccId);
+            return RetResponse.makeOKRsp(ccName);
+        }else{
+            return RetResponse.makeErrRsp("需要传入参数ccId！");
+        }
+    }
+
+
     @PostMapping("/selectByParentId")
     @ApiOperation(value = "根据父ID查询接口")
-    public RetResult<PageInfo<CourseCategoryInfo>> selectByParentId(@RequestParam String parentId) throws Exception {
-        PageHelper.startPage(0, 0);
-        List<CourseCategoryInfo> list = courseCategoryInfoService.selectByParentId(parentId);
-        PageInfo<CourseCategoryInfo> pageInfo = new PageInfo<CourseCategoryInfo>(list);
-        return RetResponse.makeOKRsp(pageInfo);
+    public RetResult<List> selectByParentId(@RequestBody JSONObject obj) throws Exception {
+        if(obj.containsKey("parentId")){
+            String parentId = obj.getString("parentId");
+            List<CourseCategoryInfo> list = courseCategoryInfoService.selectByParentId(parentId);
+            return RetResponse.makeOKRsp(list);
+        }else{
+            return RetResponse.makeErrRsp("需要传入参数parentId！");
+        }
+    }
+
+    @GetMapping("/getByParentId")
+    @ApiOperation(value = "根据父ID查询接口")
+    public RetResult<List> getByParentId(@RequestParam(required=true) String parentId) throws Exception {
+        if(!parentId.isEmpty()){
+            List<CourseCategoryInfo> list = courseCategoryInfoService.selectByParentId(parentId);
+            return RetResponse.makeOKRsp(list);
+        }else{
+            return RetResponse.makeErrRsp("入参parentId不能为空！");
+        }
     }
 
     /**
@@ -134,7 +200,7 @@ public class CourseCategoryInfoController {
      * @param size 每页条数
      * @Reutrn RetResult<PageInfo<CourseCategoryInfo>>
      */
-    @GetMapping("/list")
+    @PostMapping("/list")
     public RetResult<PageInfo<CourseCategoryInfo>> list(@RequestParam(defaultValue = "0") Integer page,
                @RequestParam(defaultValue = "0") Integer size) throws Exception {
         PageHelper.startPage(page, size);
@@ -145,15 +211,23 @@ public class CourseCategoryInfoController {
 
     /**
      * @Description: 递归查询
-     * @Reutrn RetResult<PageInfo<CourseCategoryInfo>>
+     * @Reutrn RetResult<List>
      */
-    @PostMapping("/selectListRecursion")
+    @GetMapping("/selectListRecursion")
     @ApiOperation(value = "递归查询所有接口")
-    public RetResult<PageInfo<CourseCategoryInfoVo>> selectListRecursion(@RequestParam(defaultValue = "0") Integer page,
-                @RequestParam(defaultValue = "0") Integer size) throws Exception {
-        PageHelper.startPage(page, size);
+    public RetResult<List> selectListRecursion() throws Exception {
         List<CourseCategoryInfoVo> list = courseCategoryInfoService.selectListRecursion();
-        PageInfo<CourseCategoryInfoVo> pageInfo = new PageInfo<CourseCategoryInfoVo>(list);
-        return RetResponse.makeOKRsp(pageInfo);
+        return RetResponse.makeOKRsp(list);
+    }
+
+    /**
+     * @Description: 查询所有分类
+     * @Reutrn RetResult<Map>
+     */
+    @GetMapping("/selectList")
+    @ApiOperation(value = "查询所有接口")
+    public RetResult<Map> selectList() throws Exception {
+        Map<String,String> map = courseCategoryInfoService.selectList();
+        return RetResponse.makeOKRsp(map);
     }
 }
